ai/requirements.txt
# FastAPI
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9

# AI / GraphRAG
neo4j-graphrag>=1.13.0
neo4j>=5.0.0
openai>=1.0.0

# DB
mysql-connector-python>=8.0.0

# HTTP
aiohttp>=3.9.0
requests>=2.31.0
httpx>=0.27.0

# Utils
python-dotenv>=1.0.0

# UI (optional)
gradio>=4.0.0
